requirements.txt
bcrypt==3.1.7
cffi==1.13.2
Click==7.0
dnspython==1.16.0
Flask==1.1.1
itsdangerous==1.1.0
pycparser==2.19
Werkzeug==0.16.1
scipy==1.8.0
numpy==1.22.3
pylab-sdk==1.3.2

